add a new "VisualizeData" somatic message type
This type has a variable number of vectors, a boolean vector, and a string. It can be used by
experiments to publish internal state to ach channels where it can be forwarded using achd and used
to visualize and debug the operation of the exeriments. For example, teleop/05-workspace uses it to
publish force-torque external forces and workspace position references.