change install targets to follow proper convention
When given a relative path for an install target, for example simply "include", cmake will prepend
CMAKE_INSTALL_PREFIX or CPACK_PACKAGING_INSTALL_PREFIX as appropriate for the type of installation,
automatically adding /usr/local for "make install" or /usr for debian packages. When install
commands are given absolute paths, however - like /usr/local/lib/ - they will override this
intelligent handling and will always go to that path, even when building a debian package. This is
BAD and WRONG and I'm going to have to fix everything now.